+++ web/src/features/hostdashboard/hostpricing/hooks/usePagination.js
@@ -1,0 +1,51 @@
+import { useState, useMemo } from 'react';
+
+export const usePagination = (items, itemsPerPage) => {
+  const [currentPage, setCurrentPage] = useState(1);
+  const totalPages = Math.ceil(items.length / itemsPerPage);
+
+  const paginatedItems = useMemo(() => {
+    const startIndex = (currentPage - 1) * itemsPerPage;
+    const endIndex = startIndex + itemsPerPage;
+    return items.slice(startIndex, endIndex);
+  }, [items, currentPage, itemsPerPage]);
+
+  const pageRange = useMemo(() => {
+    let startPage = currentPage - 2;
+    if (startPage < 1) startPage = 1;
+    let endPage = startPage + 4;
+    if (endPage > totalPages) {
+      endPage = totalPages;
+      startPage = Math.max(endPage - 4, 1);
+    }
+    return { startPage, endPage };
+  }, [currentPage, totalPages]);
+
+  const goToPage = (pageNumber) => {
+    if (pageNumber >= 1 && pageNumber <= totalPages) {
+      setCurrentPage(pageNumber);
+    }
+  };
+
+  const goToNextPage = () => {
+    if (currentPage < totalPages) {
+      setCurrentPage(currentPage + 1);
+    }
+  };
+
+  const goToPreviousPage = () => {
+    if (currentPage > 1) {
+      setCurrentPage(currentPage - 1);
+    }
+  };
+
+  return {
+    currentPage,
+    totalPages,
+    paginatedItems,
+    pageRange,
+    goToPage,
+    goToNextPage,
+    goToPreviousPage
+  };
+};
